# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |-
    This OpenAPI document describes APIs for the project "WASAText".

    Student: Damir Tassybayev
    Matricola: 2046612
    E-mail: tassybayev.2046612@studenti.uniroma1.it
  version: "1.0.2"
...
security:
  - BearerAuth: []
paths:
...
  /users/photo:
    get:
      tags:
        - user
      summary: Retrieves the authenticated user's photo
      description: Returns the full User object (id, name, and photo in base64)
      operationId: getMyPhoto
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User photo fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user123"
                name: "Maria"
                photo: "aGVsbG8="
    put:
      tags:
        - user
      summary: Updates the authenticated user's photo
      description: Updates the user's photo and returns the updated User object
      operationId: setMyPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: New photo (JPEG or PNG) as a base64-encoded string
        required: true
        content:
          image/png:
            schema:
              type: string
              description: PNG image file (base64 encoded)
              pattern: '^[A-Za-z0-9+/]+={0,2}$'
              minLength: 1
              maxLength: 10485760
          image/jpeg:
            schema:
              type: string
              description: JPEG image file (base64 encoded)
              pattern: '^[A-Za-z0-9+/]+={0,2}$'
              minLength: 1
              maxLength: 10485760
      responses:
        '200':
          description: Photo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user123"
                name: "Maria"
                photo: "aGVsbG8="
  /users/name:
    put:
      tags:
        - user
      summary: Updates the authenticated user's username
      description: Updates the user's username
      operationId: setMyUserName
      security:
        - BearerAuth: []
      requestBody:
        description: New username
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user123"
                name: "NewName"
                photo: "aGVsbG8="
...
  /search:
    get:
      tags:
        - chat
        - user
      summary: Searches for users by username
      description: Searches for users matching a query string.
      operationId: searchUsers
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: query
          required: true
          description: Username query parameter
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      responses:
        '200':
          description: Array of users matching the query.
          content:
            application/json:
              schema:
                type: array
                description: Array of User objects.
                minItems: 0
                maxItems: 1000
                items:
                  $ref: '#/components/schemas/User'
              example:
                - id: "user123"
                  name: "Maria"
                  photo: "aGVsbG8="
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateUserRequest:
      type: object
      description: Request schema for updating user information.
      properties:
        name:
          type: string
          description: New username.
          example: "NewName"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 3
          maxLength: 16
    User:
      type: object
      description: User schema.
      required:
        - id
        - name
        - photo
      properties:
        id:
          type: string
          description: Unique identifier of the user.
          example: "user123"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        name:
          type: string
          description: Name of the user.
          example: "Maria"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 3
          maxLength: 16
        photo:
          type: string
          description: User photo in base64 encoding.
          example: "aGVsbG8="
          pattern: '^[A-Za-z0-9+/]+={0,2}$'
          minLength: 1
          maxLength: 1000000
...